Started overhauling the add_pandas_x methods for multidimensional input.
Starting to redo the overall DataFrame structure to be somewhat more intuitive for an end user needing to surf through many different files and their respective data.  The add_pandas_x method need to be redone to handle higher dimensional data sets and hopefully allow compression along a 2 dimensional DataFrame.  Still a lot of work that needs to be done.  Both methods are pretty much totally broken for the time being.