Now able to tabulate the first N frequencies, and number of imaginary.
New frequencies method has been added that tabulates the first N frequencies and keeps track of the number of imaginary frequencies.  Can't include the full list of frequencies yet because this would exceed the number of atom lines.  Higher order frequencies are very rarely useful, so this isn't much an issue.